Makefile: move 3 header files from common/ to headers/ directory
All local header files should be put together in one directory.
Change c code to include "parsing_helpers.h" from "../common/parsing_helpers.h".
