+++ headers/xdp_stats_kern.h
@@ -4,13 +4,10 @@
 #ifndef __XDP_STATS_KERN_H
 #define __XDP_STATS_KERN_H
 
-/* Data record type 'struct datarec' is defined in common/xdp_stats_kern_user.h,
- * programs using this header must first include that file.
+/* Data record type 'struct datarec' is defined in xdp_stats_kern_user.h,
+ * and that file should always be included by default.
  */
-#ifndef __XDP_STATS_KERN_USER_H
-#warning "You forgot to #include <../common/xdp_stats_kern_user.h>"
-#include <../common/xdp_stats_kern_user.h>
-#endif
+#include "xdp_stats_kern_user.h"
 
 /* Keeps stats per (enum) xdp_action */
 struct bpf_map_def SEC("maps") xdp_stats_map = {
