+++ packet02-rewriting/Makefile
@@ -8,7 +8,8 @@ COMMON_DIR = ../common
 
 COPY_LOADER := xdp_loader
 COPY_STATS  := xdp_stats
-EXTRA_DEPS := $(COMMON_DIR)/parsing_helpers.h
+vpath %.h ../headers
+EXTRA_DEPS := parsing_helpers.h
 
 include $(COMMON_DIR)/common.mk
