+++ packet03-redirecting/Makefile
@@ -7,6 +7,7 @@ LIBBPF_DIR = ../libbpf/src/
 COMMON_DIR = ../common
 
 COPY_LOADER := xdp_loader
-EXTRA_DEPS := $(COMMON_DIR)/parsing_helpers.h
+vpath %.h ../headers
+EXTRA_DEPS := parsing_helpers.h
 
 include $(COMMON_DIR)/common.mk
