+++ packet03-redirecting/xdp_prog_kern.c
@@ -5,7 +5,7 @@
 #include "bpf_endian.h"
 
 // The parsing helper functions from the packet01 lesson have moved here
-#include "../common/parsing_helpers.h"
+#include "parsing_helpers.h"
 
 
 /* Pops the outermost VLAN tag off the packet. Returns the popped VLAN ID on
